refactor(structure): move api to hexagonal layers
