+++ src/domain/README.md
@@ -1,0 +1,1 @@
+# Domain Layer\n\nThis directory contains the domain entities, value objects, and business logic.
